LPS-97133 Update portlet preferences related to fragment entry links to point to relevant layout
For a given fragment entry link there can be up to 3 portlet preferences pointing to the company control-panel layout, the group control-panel layout and/or the layout where the fragment is placed.

If there's one portlet preferences for the group control-panel layout, it is changed to point to the layout and the others are deleted.

If there's one portlet preferences for the company control panel layout but not for the group's, also point to the layout and delete the possible one pointing to the layout. This latter one was created dynamically from the company-control-panel-layout one.