LPS-162388 Keep track of whether notification feed entry is interactive
A non-interactive entry will always be read, the body will have no link and it will have no actions.
This can happen for notification of workflow tasks that have been withdrawn, for example.

An entry can be interactive but not applicable.
The entry can then be read or unread and can have actions but the body has no link.

This can happen if a submission is withdrawn, for example. In general, if BaseUserNotificationHandler.getBody is blank.